issue #111: improved design of variable transaction count and added list of variable transactions to dashboard
