+++ frontend/src/components/home/Home.tsx
@@ -14,7 +14,7 @@ import Dashboard                                                  from './dashbo
 import Profile                                                    from './profile/Profile';
 import Settings                                                   from './settings/Settings';
 import VariableTransactionList
-                                                                  from './transactions/variable/VariableTransactionList';
+                                                                  from '../shared/transaction/variable/transactionList/VariableTransactionList';
 import CreateVariableTransaction
                                                                   from './transactions/variable/create/CreateVariableTransaction';
 import VariableTransactionsDetails
@@ -34,6 +34,8 @@ import '@ant-design/pro-table/dist/table.css';
 import '@ant-design/pro-list/dist/list.css';
 import '@ant-design/pro-card/dist/card.css';
 import UserManagement                                             from './admin/users/UserManagement';
+import VariableTransactionOverview
+                                                                  from './transactions/variable/VariableTransactionOverview';
 
 if (isDarkTheme()) {
     require('antd/dist/antd.dark.css');
@@ -130,7 +132,7 @@ class Home extends React.Component<HomeProps, HomeState> {
                         <Route path={'/transactions/variable/create'} component={CreateVariableTransaction}/>
                         <Route path={'/transactions/variable/:variableTransactionId'}
                                component={VariableTransactionsDetails}/>
-                        <Route path={'/transactions/variable/'} component={VariableTransactionList}/>
+                        <Route path={'/transactions/variable/'} component={VariableTransactionOverview}/>
                         <Route path={'/transactions/fixed/create'} component={CreateFixedTransaction}/>
                         <Route path={'/transactions/fixed/:fixedTransactionId'}
                                component={FixedTransactionDetails}/>
