+++ frontend/src/components/home/dashboard/Dashboard.tsx
@@ -2,9 +2,9 @@ import {AppState}                         from '../../../store/reducers/root.red
 import {bindActionCreators, Dispatch}     from 'redux';
 import {connect}                          from 'react-redux';
 import {WithTranslation, withTranslation} from 'react-i18next';
-import React                              from 'react';
-import {UserReducerState}                 from '../../../store/reducers/user.reducers';
-import {PageContainer}                    from '@ant-design/pro-layout';
+import React           from 'react';
+import {UserState}     from '../../../store/reducers/user.reducers';
+import {PageContainer} from '@ant-design/pro-layout';
 import BalanceHistoryChart                from '../../shared/statistics/balance/BalanceHistoryChart';
 import CategoryDistributionChart
                                           from '../../shared/statistics/categoryDistribution/CategoryDistributionChart';
@@ -12,7 +12,7 @@ import {Col, Row}                         from 'antd';
 import VariableTransactionCountHistoryChart
                                           from '../../shared/statistics/variableTransaction/variableTransactionCount/VariableTransactionCountChart';
 
-interface DashboardProps extends WithTranslation<'default'>, UserReducerState {
+interface DashboardProps extends WithTranslation<'default'>, UserState {
 }
 
 interface DashboardState {
@@ -29,10 +29,10 @@ class Dashboard extends React.Component<DashboardProps, DashboardState> {
                     </Col>
                 </Row>
                 <Row gutter={[16, 16]}>
-                    <Col span={12}>
+                    <Col span={24} md={{span: 12}}>
                         <CategoryDistributionChart/>
                     </Col>
-                    <Col span={12}>
+                    <Col span={24}  md={{span: 12}}>
                         <VariableTransactionCountHistoryChart/>
                     </Col>
                 </Row>
